First dump of content from original collection
This includes all of the source documentation of the AP Computer Science
A Curriculum from the instructor Sharepoint site, minus all PDF files
that are saved transforms of the original Word (.docx) or PowerPoint
(.pptx) files.

This also includes first versions of the project README.md and
LICENSE.md files.