Move String cast to note
Since String conversion is not a cast, it has been moved a note.  I changed the cast to use implicit String conversion in concatenation for two reasons. 1) the AP CS A Java subset states down in note 5: "String concatenation + is part of the AP Java subset.  Students are expected to know that concatenation converts numbers to strings and invokes toString on objects. And 2) the "" + is consistent with how the text book "Building Java Programs" converts numbers to Strings.